feat: move HeroBannerRating component to folder and add the test
