+++ src/components/molecules/HeroBanner.tsx
@@ -5,7 +5,7 @@ import { Link } from "react-router-dom";
 import HeroBannerDescription from "../atoms/HeroBannerDescription/HeroBannerDescription";
 import HeroBannerTags from "../atoms/HeroBannerTags";
 import HeroBannerTitle from "../atoms/HeroBannerTitle";
-import HeroBannerRating from "../atoms/HeroBannerRating";
+import HeroBannerRating from "../atoms/HeroBannerRatings/HeroBannerRating";
 const image_url = import.meta.env.VITE_IMAGE_URL;
 import { Movie } from "../../types/Movie";
